feat: add input validations to file manager CLI

+++ nodejs-revition/nodejs-cli-application/file-manager-cli/src/commands/copyCommands.js
@@ -2,14 +2,18 @@ const path = require("path");
 const fs = require("fs");
 function copyFile(filename){
     const oldPath = path.join(__dirname,"..","..","workspace",filename);
-    const destination = path.join(__dirname,"..","..", "backup",filename);
-    console.log(destination);
-    fs.copyFile(oldPath,destination,(err) => {
-        if(err){
-            console.log(err);
-            // return;
-        }
-        console.log("✓ File copied successfully.");
-    })
+    const destination = path.join(__dirname,"..","..","backups",filename);
+    if(fs.existsSync(oldPath)){
+        fs.copyFile(oldPath,destination,(err) => {
+            if(err){
+                throw err;
+            }
+            console.log("✓ File copied successfully.");
+        });
+    }
+    else{
+        console.log("file does not exits ❌");
+    }
+
 }
 module.exports = copyFile;
