+++ nodejs-revition/nodejs-cli-application/file-manager-cli/src/commands/renameCommand.js
@@ -4,14 +4,18 @@ const path = require('path');
 function renameFile(oldfileName,newfileName){
     const oldfilePath = path.join(__dirname,"..","..", "workspace",oldfileName);
     const newfilePath = path.join(__dirname,"..","..","workspace",newfileName);
-
-    fs.rename(oldfilePath,newfilePath, (err) => {
-        if(err){
+    if(fs.existsSync(oldfilePath)){
+        fs.rename(oldfilePath,newfilePath, (err) => {
+           if(err){
             console.log(err);
             return;
         }
     });
-    console.log("✓ File renamed successfully.");
+    console.log("✓ File renamed successfully.");   
+    }
+    else{
+        console.log("file does not exits ❌");
+    }
 }
 
 module.exports = renameFile;
