+++ nodejs-revition/nodejs-cli-application/file-manager-cli/src/commands/rmdirCommand.js
@@ -4,12 +4,17 @@ const path = require('path');
 
 function removeDirectory(foldername){
     const filePath = path.join(__dirname,"..","..","workspace",foldername);
-    fs.rm(filePath,{recursive:true},(err) => {
+    if(fs.existsSync(filePath)){
+        fs.rm(filePath,{recursive:true},(err) => {
         if(err){
             throw err;
         }
         console.log("✓ Directory deleted successfully.");
-    })
+    })   
+    }
+    else{
+        console.log("Directory does not exist ❌")
+    }
 }
 
 module.exports = removeDirectory;
